use big decimal for money related fields

+++ code/v1/backend/src/main/java/com/thy/minimalist/budgetingIncomesExpenses/model/Record.java
@@ -34,10 +34,10 @@ public class Record {
     private BigDecimal actualAmount;
 
     @Column(name = "expected_amount", precision = 15, scale = 2)
-    private Long expectedAmount;
+    private BigDecimal expectedAmount;
 
     @Column(name = "charges", precision = 15, scale = 2)
-    private Long charge;
+    private BigDecimal charge;
 
     @NotNull
     @ManyToOne
